Another large shakeup in the code.
Lots and lots of temporary debug code, mostly to monitor memory usage. So far so good.

Might merge the the handshake code from "GenericClient" into the "Server" class.

The TLS server is a work in progress.
It does have hickups, but as yet I was unable to identify the cause.
There is very little in the way of error detection for the streams. I will need to read up on how to do that properly.
The TLS handshake is in blocking mode, which needs to be fixed.